partition rpc into multiple files
