+++ tests/unittest/executor/test_rpc.py
@@ -1,4 +1,3 @@
-import threading
 import time
 
 import pytest
@@ -7,53 +6,129 @@ from tensorrt_llm.executor.rpc import (RPCCancelled, RPCClient, RPCError,
                                        RPCServer, RPCTimeout)
 
 
-def test_rpc_server_basics():
+class RpcServerWrapper(RPCServer):
 
-    class App:
+    def __init__(self, *args, addr: str, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.addr = addr
 
-        def hello(self):
-            print("hello")
+    def __enter__(self):
+        self.bind(self.addr)
+        self.start()
+        return self
 
-    server = RPCServer(App())
-    print("bind")
-    server.bind("ipc:///tmp/rpc_test")
-    print("start")
-    server.start()
-    print("sleep")
-
-    time.sleep(1)
-    print("shutdown")
-    server.shutdown()
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.shutdown()
 
 
-def test_rpc_client_context_manager():
+class TestRpcBasics:
 
-    class App:
+    def test_rpc_server_basics(self):
 
-        def hello(self):
-            print("hello")
+        class App:
 
-    with RPCServer(App()) as server:
-        server.bind("ipc:///tmp/rpc_test")
-        server.start()
-        time.sleep(1)
+            def hello(self):
+                print("hello")
 
+        with RpcServerWrapper(App(), addr="ipc:///tmp/rpc_test") as server:
+            pass
 
-def test_rpc_hello_without_arg():
+    def test_remote_call_without_arg(self):
 
-    class App:
+        class App:
 
-        def hello(self):
-            print("hello")
-            return "world"
+            def hello(self):
+                print("hello")
+                return "world"
 
-    with RPCServer(App()) as server:
-        server.bind("ipc:///tmp/rpc_test")
-        server.start()
-        time.sleep(0.1)
-        client = RPCClient("ipc:///tmp/rpc_test")
-        ret = client.hello()  # sync call
-        assert ret == "world"
+        with RpcServerWrapper(App(), addr="ipc:///tmp/rpc_test") as server:
+            client = RPCClient("ipc:///tmp/rpc_test")
+            ret = client.hello()  # sync call
+            assert ret == "world"
+
+    def test_remote_call_with_args(self):
+
+        class App:
+
+            def hello(self, name: str, location: str):
+                print("hello")
+                return f"hello {name} from {location}"
+
+        with RpcServerWrapper(App(), addr="ipc:///tmp/rpc_test") as server:
+            client = RPCClient("ipc:///tmp/rpc_test")
+            ret = client.hello("app", "Marvel")
+            assert ret == "hello app from Marvel"
+
+    def test_remote_call_with_kwargs(self):
+
+        class App:
+
+            def hello(self, name: str, location: str):
+                print("hello")
+                return f"hello {name} from {location}"
+
+        with RpcServerWrapper(App(), addr="ipc:///tmp/rpc_test") as server:
+            client = RPCClient("ipc:///tmp/rpc_test")
+            ret = client.hello(name="app", location="Marvel")
+            assert ret == "hello app from Marvel"
+
+    def test_remote_call_with_args_and_kwargs(self):
+
+        class App:
+
+            def hello(self, name: str, location: str):
+                print("hello")
+                return f"hello {name} from {location}"
+
+        with RpcServerWrapper(App(), addr="ipc:///tmp/rpc_test") as server:
+            client = RPCClient("ipc:///tmp/rpc_test")
+            ret = client.hello(name="app", location="Marvel")
+            assert ret == "hello app from Marvel"
+
+    def test_rpc_server_address(self):
+
+        class App:
+            pass
+
+        with RpcServerWrapper(App(), addr="ipc:///tmp/rpc_test") as server:
+            assert server.address == "ipc:///tmp/rpc_test"
+
+    def test_rpc_with_error(self):
+
+        class App:
+
+            def hello(self):
+                raise ValueError("hello")
+
+        with RpcServerWrapper(App(),
+                              addr="ipc:///tmp/rpc_test_error") as server:
+            client = RPCClient("ipc:///tmp/rpc_test_error")
+            with pytest.raises(RPCError):
+                client.hello()
+
+    def test_rpc_without_wait_response(self):
+
+        class App:
+
+            def __init__(self):
+                self.task_submitted = False
+
+            def send_task(self) -> None:
+                # Just submit the task and return immediately
+                # The result is not important
+                self.task_submitted = True
+                return None
+
+            def get_task_submitted(self) -> bool:
+                return self.task_submitted
+
+        with RpcServerWrapper(App(),
+                              addr="ipc:///tmp/rpc_test_no_wait") as server:
+            client = RPCClient("ipc:///tmp/rpc_test_no_wait")
+            client.send_task(__rpc_need_response=False)
+            time.sleep(
+                0.1)  # wait for some time to make sure the task is submitted
+            assert client.get_task_submitted()
 
 
 class TestRpcError:
@@ -138,7 +213,7 @@ class TestRpcError:
             with pytest.raises(RPCCancelled):
                 future.result()
 
-        time.sleep(1)
+        time.sleep(5)
 
         client.close()
 
@@ -211,80 +286,6 @@ def test_rpc_shutdown_server():
     time.sleep(5)  # the server dispatcher thread need some time to quit
 
 
-def test_rpc_hello_with_arg():
-
-    class App:
-
-        def hello(self, name: str, location: str):
-            print("hello")
-            return f"hello {name} from {location}"
-
-    with RPCServer(App()) as server:
-        server.bind("ipc:///tmp/rpc_test")
-        server.start()
-        time.sleep(0.1)
-        client = RPCClient("ipc:///tmp/rpc_test")
-        ret = client.hello("app", location="Marvel")  # sync call
-        assert ret == "hello app from Marvel"
-
-
-def test_rpc_server_address():
-
-    class App:
-
-        def hello(self):
-            print("hello")
-            return "world"
-
-    with RPCServer(App()) as server:
-        server.bind("ipc:///tmp/rpc_test")
-        server.start()
-        time.sleep(0.1)
-        assert server.address == "ipc:///tmp/rpc_test"
-
-
-def test_rpc_with_error():
-
-    class App:
-
-        def hello(self):
-            raise ValueError("hello")
-
-    with RPCServer(App()) as server:
-        server.bind("ipc:///tmp/rpc_test_error")
-        server.start()
-        time.sleep(0.1)
-        client = RPCClient("ipc:///tmp/rpc_test_error")
-        with pytest.raises(RPCError):
-            client.hello()
-
-
-def test_rpc_without_wait_response():
-
-    class App:
-
-        def __init__(self):
-            self.task_submitted = False
-
-        def send_task(self) -> None:
-            # Just submit the task and return immediately
-            # The result is not important
-            self.task_submitted = True
-            return None
-
-        def get_task_submitted(self) -> bool:
-            return self.task_submitted
-
-    with RPCServer(App()) as server:
-        server.bind("ipc:///tmp/rpc_test_no_wait")
-        server.start()
-        time.sleep(0.1)
-        client = RPCClient("ipc:///tmp/rpc_test_no_wait")
-        client.send_task(__rpc_need_response=False)
-        time.sleep(0.1)  # wait for some time to make sure the task is submitted
-        assert client.get_task_submitted()
-
-
 def test_rpc_without_response_performance():
     # At any circumstances, the RPC call without response should be faster than the one with response
     class App:
@@ -339,7 +340,7 @@ def test_rpc_benchmark(async_run_task: bool, use_ipc_addr: bool):
         client = RPCClient(server.address)
 
         time_start = time.time()
-        for i in range(10000):
+        for i in range(100):
             ret = client.cal(i, __rpc_timeout=10)  # sync call
             assert ret == i * 2, f"{ret} != {i * 2}"
         time_end = time.time()
@@ -371,9 +372,8 @@ def test_rpc_timeout(use_async: bool):
 
         # Test that a short timeout causes RPCTimeout exception
         with pytest.raises(RPCTimeout) as exc_info:
+            import asyncio
             if use_async:
-                # Test async call with timeout
-                import asyncio
 
                 async def test_async_timeout():
                     return await client.call_async('slow_operation',
@@ -405,27 +405,14 @@ def test_rpc_timeout(use_async: bool):
         client.close()
 
 
-class RpcServerWrapper(RPCServer):
-
-    def __init__(self, *args, addr: str, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.addr = addr
-
-    def __enter__(self):
-        self.bind(self.addr)
-        self.start()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.shutdown()
-
-
 class TestRpcShutdown:
 
     def test_duplicate_shutdown(self):
 
         class App:
-            pass
+
+            def quick_task(self, task_id: int):
+                return f"quick_task_{task_id}"
 
         with RpcServerWrapper(App(),
                               addr="ipc:///tmp/rpc_test_shutdown") as server:
@@ -441,124 +428,23 @@ class TestRpcShutdown:
 
         class App:
 
-            def foo(self):
+            def foo(self, delay: int):
+                time.sleep(delay)
                 return "foo"
 
-        with RpcServerWrapper(App(),
-                              addr="ipc:///tmp/rpc_test_shutdown") as server:
-            client = RPCClient("ipc:///tmp/rpc_test_shutdown")
-            assert client.foo() == "foo"
-
-            server.shutdown()
-            time.sleep(2)  # wait for the server to shutdown
-
-            with pytest.raises(RPCServerShuttingDown):
-                client.foo()
-
-    def test_submit_request_after_shutdown(self):
-
-        class App:
-
-            def __init__(self):
-                self.completed_requests = []
-                self.lock = threading.Lock()
-
-            def quick_task(self, task_id):
-                with self.lock:
-                    self.completed_requests.append(task_id)
-                return f"quick_task_{task_id}_completed"
-
-        app = App()
-        with RPCServer(app) as server:
-            server.bind("ipc:///tmp/rpc_test_shutdown")
-            server.start()
-            time.sleep(0.1)
-            client = RPCClient("ipc:///tmp/rpc_test_shutdown")
-            client.quick_task(1)
-            server.shutdown()
-            time.sleep(2)  # wait for the server to shutdown
-            with pytest.raises(RPCServerShuttingDown):
-                client.quick_task(2)
-
-
-def test_enhanced_shutdown_mechanism():
-    """Test enhanced shutdown mechanism that refuses new requests and waits for pending ones."""
-
-    class App:
-
-        def __init__(self):
-            self.completed_requests = []
-            self.lock = threading.Lock()
-
-        def quick_task(self, task_id):
-            """A quick task for testing."""
-            with self.lock:
-                self.completed_requests.append(task_id)
-            return f"quick_task_{task_id}_completed"
-
-        def slow_task(self, task_id, duration=1.0):
-            """A slow task that takes some time to complete."""
-            time.sleep(duration)
-            with self.lock:
-                self.completed_requests.append(task_id)
-            return f"slow_task_{task_id}_completed"
-
-    app = App()
-
-    with RPCServer(app) as server:
-        server.bind("ipc:///tmp/rpc_test_enhanced_shutdown")
+        server = RPCServer(App())
+        server.bind("ipc:///tmp/rpc_test_shutdown")
         server.start()
+
         time.sleep(0.1)
+        client = RPCClient("ipc:///tmp/rpc_test_shutdown")
+        # This task should be continued after server shutdown
+        res = client.foo(10, __rpc_timeout=12, __rpc_mode="future")
 
-        client = RPCClient("ipc:///tmp/rpc_test_enhanced_shutdown")
+        # The shutdown will block until all pending requests are finished
+        server.shutdown()
 
-        # Submit some quick requests that should complete
-        quick_future1 = client.quick_task.call_future(1)
-        quick_future2 = client.quick_task.call_future(2)
-
-        # Submit a slow request that should still complete during shutdown
-        slow_future = client.slow_task.call_future(10, duration=0.5)
-
-        # Wait a bit to ensure requests are queued
-        time.sleep(0.1)
-
-        # Start shutdown in a separate thread to test concurrent behavior
-        def shutdown_server():
-            time.sleep(0.2)  # Let pending requests start processing
-            server.shutdown()
-
-        shutdown_thread = threading.Thread(target=shutdown_server)
-        shutdown_thread.start()
-
-        # Try to submit a new request after shutdown has started
-        # This should eventually fail with RPCServerShuttingDown
-        time.sleep(0.3)  # Wait for shutdown to start
-
-        try:
-            client.quick_task(999)
-            pytest.fail("Expected RPCServerShuttingDown exception")
-        except RPCServerShuttingDown as e:
-            assert "Server is shutting down" in str(e)
-            print(f"✓ New request properly refused during shutdown: {e}")
-        except Exception as e:
-            # Might get other connection errors which is also acceptable
-            print(f"✓ New request failed as expected during shutdown: {e}")
-
-        # Wait for shutdown to complete
-        shutdown_thread.join(timeout=5.0)
-
-        # Check that all pending requests completed
-        assert quick_future1.result() == "quick_task_1_completed"
-        assert quick_future2.result() == "quick_task_2_completed"
-        assert slow_future.result() == "slow_task_10_completed"
-
-        # Verify all tasks were completed
-        with app.lock:
-            assert 1 in app.completed_requests
-            assert 2 in app.completed_requests
-            assert 10 in app.completed_requests
-
-        print("✓ Enhanced shutdown mechanism test passed")
+        assert res.result() == "foo"
 
 
 if __name__ == "__main__":
